+++ library/src/androidTest/java/com/twilio/video/util/FrameCountRenderer.java
@@ -5,10 +5,7 @@ import com.twilio.video.VideoRenderer;
 
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicReference;
-
-import static junit.framework.TestCase.assertNotNull;
 
 public class FrameCountRenderer implements VideoRenderer {
     private final AtomicReference<CountDownLatch> frameArrived =
@@ -21,9 +18,6 @@ public class FrameCountRenderer implements VideoRenderer {
 
     @Override
     public void renderFrame(I420Frame frame) {
-        assertNotNull(frame);
-        assertNotNull(frame.yuvPlanes);
-        assertNotNull(frame.yuvStrides);
         frameCount++;
         frameArrived.get().countDown();
         frame.release();