+++ library/src/main/java/com/twilio/video/I420Frame.java
@@ -1,9 +1,25 @@
 package com.twilio.video;
 
+import android.support.annotation.Nullable;
+
 import java.nio.ByteBuffer;
 
 /**
- * A YUV frame in the I420 format.
+ * A YUV frame in the I420 format. A frame can be represented as a {@link ByteBuffer} array of
+ * Y, U, and V pixel data with an array of strides for each plane or as a texture.
+ *
+ * <p>
+ *
+ * When a frame is represented as a texture, {@link #textureId} will be set to a positive
+ * non zero value with {@link #yuvPlanes} and {@link #yuvStrides} set to {@code null}. The YUV
+ * data can be extracted from the texture using an instance of {@link org.webrtc.YuvConverter}
+ * and the {@link #samplingMatrix}.
+ *
+ * <p>
+ *
+ * When a frame is represented as an array of {@link ByteBuffer}, {@link #textureId} will be
+ * 0, {@link #yuvPlanes} contains the YUV pixel data, and {@link #yuvStrides} contains each
+ * plane's stride.
  */
 public class I420Frame {
     /** Width of frame */
@@ -11,32 +27,44 @@ public class I420Frame {
     /** Height of frame */
     public final int height;
     /**
-     * Array of strides for each plane.
+     * Array of strides for each plane. This field is {@code null} when the frame is represented
+     * as a texture.
      *
-     * yuvStrides[0] - Y stride.
-     * yuvStrides[1] - U stride.
-     * yuvStrides[2] - V stride.
+     * <p>
+     *
+     * yuvStrides[0] - Y stride. <br/>
+     * yuvStrides[1] - U stride. <br/>
+     * yuvStrides[2] - V stride. <br/>
      */
-    public final int[] yuvStrides;
+    @Nullable public final int[] yuvStrides;
     /**
-     * Array of pixel data for each plane.
+     * Array of pixel data for each plane. This field is {@code null} when the frame is represented
+     * as a texture.
      *
-     * yuvPlanes[0] - Y pixel data.
-     * yuvPlanes[1] - U pixel data.
-     * yuvPlanes[2] - V pixel data.
+     * <p>
+     *
+     * yuvPlanes[0] - Y pixel data. <br/>
+     * yuvPlanes[1] - U pixel data. <br/>
+     * yuvPlanes[2] - V pixel data. <br/>
      */
-    public final ByteBuffer[] yuvPlanes;
+    @Nullable public final ByteBuffer[] yuvPlanes;
     /** The degree that the frame must be rotated clockwise to be rendered correctly. */
     public int rotationDegree;
-    final org.webrtc.VideoRenderer.I420Frame webRtcI420Frame;
-    long nativeFramePointer;
-    /*
+    /**
+     * Id of the texture the frame is bound to. This field is zero when the frame is
+     * represented as buffer.
+     */
+    public int textureId;
+   /**
      * Matrix that transforms standard coordinates to their proper sampling locations in
      * the texture. This transform compensates for any properties of the video source that
      * cause it to appear different from a normalized texture. This matrix does not take
-     * |rotationDegree| into account.
+     * {@link #rotationDegree} into account.
      */
-    private final float[] samplingMatrix;
+    public final float[] samplingMatrix;
+
+    final org.webrtc.VideoRenderer.I420Frame webRtcI420Frame;
+    long nativeFramePointer;
 
     I420Frame(org.webrtc.VideoRenderer.I420Frame webRtcI420Frame) {
         this.width = webRtcI420Frame.width;
@@ -46,21 +74,13 @@ public class I420Frame {
         this.rotationDegree = webRtcI420Frame.rotationDegree;
         this.webRtcI420Frame = webRtcI420Frame;
         this.nativeFramePointer = webRtcI420Frame.nativeFramePointer;
+        this.textureId = webRtcI420Frame.textureId;
+        this.samplingMatrix = webRtcI420Frame.samplingMatrix;
+
         if (rotationDegree % 90 != 0) {
             throw new IllegalArgumentException("Rotation degree not multiple of 90: " +
                     rotationDegree);
         }
-        /*
-         * The convention in WebRTC is that the first element in a ByteBuffer corresponds to the
-         * top-left corner of the image, but in glTexImage2D() the first element corresponds to the
-         * bottom-left corner. This discrepancy is corrected by setting a vertical flip as sampling
-         * matrix.
-         */
-        samplingMatrix = new float[] {
-                1,  0, 0, 0,
-                0, -1, 0, 0,
-                0,  0, 1, 0,
-                0,  1, 0, 1};
     }
 
     /**